♻️ addTimelineをchat.tsではなくtimeline.tsで定義するように
