+++ src/atoms/timeline.ts
@@ -1,4 +1,6 @@
 import { atom, useAtom } from 'jotai';
+import { useCallback } from 'react';
+import { useNotification } from '@/lib/notification';
 
 export type timeline = {
   author: string;
@@ -7,8 +9,19 @@ export type timeline = {
 };
 
 export const TimelineAtom = atom<timeline[]>([]);
+export const MESSAGE_LIMIT = 3;
 
 export const useTimelineAtom = () => {
-  const [atom] = useAtom(TimelineAtom);
-  return atom;
+  const { notify } = useNotification();
+  const [timeline, setTimeline] = useAtom(TimelineAtom);
+
+  const addTimeline = useCallback(
+    (data: timeline) => {
+      setTimeline((timeline) => [data, ...timeline].filter((_, i) => i < 4));
+      notify();
+    },
+    [notify, setTimeline],
+  );
+
+  return { timeline, addTimeline };
 };
